Renamed directories, cleanup, and removed query functions
