+++ docker-compose.yml
@@ -3,9 +3,9 @@ version: '3.7'
 services:
   web:
     build: ./
-    command: uvicorn app.main:app --reload --workers 1 --host 0.0.0.0 --port 8000
+    command: uvicorn policyguru.main:app --reload --workers 1 --host 0.0.0.0 --port 8000
     volumes:
-      - ./src/:/usr/src/app/
+      - ./src/:/usr/src/policyguru/
     ports:
       - 8002:8000
