+++ docs/contributing/running-locally.md
@@ -1,6 +1,6 @@
-# Running locally
+# Running the API locally
 
-## Local FastAPI
+## Option 1: Local FastAPI
 
 We set up a FastAPI option for local testing and development purposes - particularly for testing out the UI.
 
@@ -18,36 +18,15 @@ pip3 install -r requirements-dev.txt
 make run
 ```
 
-## Invoking Lambdas locally
+## Option 2: Docker
 
-### Option 1: Leverage PyInvoke command
+You can build a Docker container that hosts this locally. This is hosted on the **8002** local port.
 
-Alternatively, you can just run the [PyInvoke](http://www.pyinvoke.org/) commands that will run all of the above:
-
-```bash
-# list available PyInvoke commands
-invoke -l
-invoke integration.sam-invoke
-```
-
-### Option 2: Run individual commands
-
-First you will need to build the serverless application using the `sam build --use-container`. This command gathers the build artifacts of your application's dependencies and places them in the proper format and location for next steps, such as locally testing, packaging, and deploying.
-
-* Run the build command:
+* Build the docker container with docker-compose:
 
 ```bash
-sam build --use-container
+make run-docker
 ```
 
-* Run functions locally and invoke them with the `sam local invoke` command.
-
-```bash
-sam local invoke WritePolicyFunction --event events/write-policy-mock.json
-sam local invoke ScanPolicyFunction --event events/scan-policy-mock.json
-sam local invoke QueryActionsFunction --event events/query-actions-mock.json
-sam local invoke QueryResourcesFunction --event events/query-resources-mock.json
-sam local invoke QueryConditionsFunction --event events/query-conditions-mock.json
-```
-
+* You can access the docs at the **8002** port: http://localhost:8002/docs
