--- events/query-actions-mock.json
@@ -1,6 +1,0 @@
-{
-  "queryStringParameters": {
-      "service": "s3",
-      "name": "GetObject"
-  }
-}
