--- events/query-conditions-mock.json
@@ -1,5 +1,0 @@
-{
-   "queryStringParameters": {
-      "service": "secretsmanager"
-  }
-}
