+++ tasks.py
@@ -126,30 +126,6 @@ def run_sam_invoke(c):
         c.run('sam build --use-container')
         c.run('sam local invoke WritePolicyFunction --event events/write-policy-mock.json')
         c.run('sam local invoke ScanPolicyFunction --event events/scan-policy-mock.json')
-        c.run('sam local invoke QueryActionsFunction --event events/query-actions-mock.json')
-        c.run('sam local invoke QueryResourcesFunction --event events/query-resources-mock.json')
-        c.run('sam local invoke QueryConditionsFunction --event events/query-conditions-mock.json')
-    except UnexpectedExit as u_e:
-        logger.critical(f"FAIL! UnexpectedExit: {u_e}")
-        sys.exit(1)
-    except Failure as f_e:
-        logger.critical(f"FAIL: Failure: {f_e}")
-        sys.exit(1)
-
-
-# INTEGRATION TESTS - Run the lambdas directly to make sure that the __main__ works properly
-@task
-def run_lambdas_directly(c):
-    """Integration testing: Running the Lambdas directly to validate that the __main__ works properly with
-    its dummy values"""
-    c.run('echo "Integration testing: Running the Lambdas directly to validate that the __main__ works properly with '
-          'its dummy values"')
-    try:
-        c.run('python3 lambdas/scan_policy/app.py')
-        c.run('python3 lambdas/write_policy/app.py')
-        c.run('python3 lambdas/query_actions/app.py')
-        c.run('python3 lambdas/query_resources/app.py')
-        c.run('python3 lambdas/query_conditions/app.py')
     except UnexpectedExit as u_e:
         logger.critical(f"FAIL! UnexpectedExit: {u_e}")
         sys.exit(1)
@@ -171,4 +147,3 @@ test.add_task(security_scan, 'security')
 
 integration.add_task(run_sam_invoke, 'sam-invoke')
 integration.add_task(run_sam_validate, 'sam-validate')
-integration.add_task(run_lambdas_directly, 'run-lambdas-directly')
