+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 from starlette.testclient import TestClient
 
-from app.main import app
+from policyguru.main import app
 
 
 @pytest.fixture(scope="module")
