+++ tests/test_main.py
@@ -1,7 +1,7 @@
 import unittest
 from starlette.testclient import TestClient
 
-from app.main import app
+from policyguru.main import app
 
 client = TestClient(app)
